CORRECT header and footer components, import them properly into App.js, and paste in Google Fonts imports
